+++ core/src/main/kotlin/no/elg/hex/util/IslandProgress.kt
@@ -3,31 +3,34 @@ package no.elg.hex.util
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.Preferences
 import no.elg.hex.Hex
+import no.elg.hex.model.FastIslandMetadata
+import no.elg.hex.preview.PreviewModifier
 import no.elg.hex.screens.PreviewIslandScreen
 
 val islandPreferences: Preferences by lazy { Gdx.app.getPreferences("island") }
-fun getPrefName(id: Int, preview: Boolean) = "$id${if (preview) "-preview" else ""}"
-fun getProgress(id: Int, preview: Boolean = false): String? {
+fun getPrefName(id: Int, preview: Boolean = false) = "$id${if (preview) "-preview" else ""}"
+fun getIslandProgress(id: Int, preview: Boolean = false): String? {
   if (Hex.args.mapEditor) {
     return null
   }
-  val pref = getPrefName(id, preview)
+  val pref = getPrefName(id)
   return islandPreferences.getString(pref, null)
 }
 
-fun PreviewIslandScreen.saveProgress() {
+fun PreviewIslandScreen.saveIslandProgress() {
   Gdx.app.debug("IS PROGRESS", "Saving progress of island ${metadata.id}")
   island.select(null)
-  islandPreferences.putString(getPrefName(metadata.id, false), island.createDto().serialize())
+  islandPreferences.putString(getPrefName(metadata.id), island.createDto().serialize())
   islandPreferences.flush()
   if (!Hex.paused) {
     Hex.assets.islandPreviews.updateSelectPreview(metadata, island)
   }
 }
 
-fun clearIslandProgress(id: Int) {
-  Gdx.app.debug("IS PROGRESS", "Clearing progress of island $id")
-  islandPreferences.remove(getPrefName(id, false))
-  islandPreferences.remove(getPrefName(id, true))
+fun clearIslandProgress(metadata: FastIslandMetadata) {
+  Gdx.app.debug("IS PROGRESS", "Clearing progress of island ${metadata.id}")
+  metadata.modifier = PreviewModifier.NOTHING
+  islandPreferences.remove(getPrefName(metadata.id))
+  islandPreferences.remove(getPrefName(metadata.id, true))
   islandPreferences.flush()
 }